+++ tests/_support/Step/Vpn/AdmEditRealtyObject.php
@@ -29,11 +29,13 @@ class AdmEditRealtyObject extends \VpnTester
         $I->waitForElement(AdmRealtyObject::$objStatus1);
         $I->click(AdmRealtyObject::$objStatus1);
         $I->click(AdmRealtyObject::$regionField);
+        $I->fillField(AdmRealtyObject::$typeRegion, Flat::region);
         $I->click(AdmRealtyObject::$region0);
         $I->click(AdmRealtyObject::$cityField);
-        $I->click(AdmRealtyObject::$city4);
+        $I->fillField(AdmRealtyObject::$typeCity, Flat::city);
+        $I->click(AdmRealtyObject::$city0);
         $I->click(AdmRealtyObject::$districtField);
-        $I->click(AdmRealtyObject::$district24);
+        $I->click(AdmRealtyObject::$district22);
         $I->click(AdmRealtyObject::$street);
         $I->fillField(AdmRealtyObject::$typeStreet, Flat::editStreet);
         $I->click(AdmRealtyObject::$chooseStreet);
@@ -133,9 +135,11 @@ class AdmEditRealtyObject extends \VpnTester
         $I->waitForElement(AdmRealtyObject::$objStatus1);
         $I->click(AdmRealtyObject::$objStatus1);
         $I->click(AdmRealtyObject::$regionField);
+        $I->fillField(AdmRealtyObject::$typeRegion, House::region);
         $I->click(AdmRealtyObject::$region0);
         $I->click(AdmRealtyObject::$cityField);
-        $I->click(AdmRealtyObject::$city4);
+        $I->fillField(AdmRealtyObject::$typeCity, House::city);
+        $I->click(AdmRealtyObject::$city0);
         $I->click(AdmRealtyObject::$districtField);
         $I->click(AdmRealtyObject::$district3);
         $I->click(AdmRealtyObject::$street);
@@ -253,9 +257,11 @@ class AdmEditRealtyObject extends \VpnTester
         $I->waitForElement(AdmRealtyObject::$objStatus1);
         $I->click(AdmRealtyObject::$objStatus1);
         $I->click(AdmRealtyObject::$regionField);
+        $I->fillField(AdmRealtyObject::$typeRegion, Parcel::region);
         $I->click(AdmRealtyObject::$region0);
         $I->click(AdmRealtyObject::$cityField);
-        $I->click(AdmRealtyObject::$city4);
+        $I->fillField(AdmRealtyObject::$typeCity, Parcel::city);
+        $I->click(AdmRealtyObject::$city0);
         $I->click(AdmRealtyObject::$districtField);
         $I->click(AdmRealtyObject::$district7);
         $I->click(AdmRealtyObject::$street);
@@ -344,9 +350,11 @@ class AdmEditRealtyObject extends \VpnTester
         $I->waitForElement(AdmRealtyObject::$objStatus1);
         $I->click(AdmRealtyObject::$objStatus1);
         $I->click(AdmRealtyObject::$regionField);
+        $I->fillField(AdmRealtyObject::$typeRegion, Commercial::region);
         $I->click(AdmRealtyObject::$region0);
         $I->click(AdmRealtyObject::$cityField);
-        $I->click(AdmRealtyObject::$city4);
+        $I->fillField(AdmRealtyObject::$typeCity, Commercial::city);
+        $I->click(AdmRealtyObject::$city0);
         $I->click(AdmRealtyObject::$districtField);
         $I->click(AdmRealtyObject::$district3);
         $I->click(AdmRealtyObject::$street);
